LibStatsD should be a .net 3.5 dll so it can be used by components old and new...
Sadly however, the justeat.testing dll was changed to run on .net 4.0.

Why is justeat.testing a .net 4 dll?? i dont really see a reason... apparently for frontend which runs on .net4 now, however it should have still been able to use it with .net 3.5 without issue.

So still looking for a good reason, if there isn't one, then it should be downgraded back to .net 3.5 and then i will be happy to change this back to .net 3.5 so the MAJORITY of components can use it.

So for now, if you want to use this, you need to be on .net 4.