Enabling Batch Metrics to be sent via lib statsd - two steps 1: Introduce IEnumerable Send Method to libs stats, iterates through applies newlines to separate metrics (real ballache) then convert to bytes.
2. Improve sending mechanism across the board, have utilized code from NanoTube, which is what the guys who made PerfTap built, it maximizes udp packet size before sending and makes our udp client asynchronous and uses object pooling to do this.
We convert all metrics being sent as single udp packets to an Innumerable string and go through the same send process.  So one pipeline.

I have upgraded their code as it results in bad lines being sent to statsd, might move the changes back downstream to them, but they have been inactive for a while, should mention the sample cod from which libs statsd message formatting also has been updated as you cannot just append new lines to the ends of metrics as it was doing, as again, this is a bad line to statsd, so all new line appendage happens at the pre-convert to bytes part in the packet builder now... and ONLY if its required... we should see a massive reduction of badlines now when Perftap uses it.
also note there is a possible bug within their try catching which misses bugs like cant connect to server, but if this hits us i think we can easily improve the code, just don't want to at this stage.
FxCop was gassing on about using GC.SupressFinalize and overriding Dispose with a bool gas, so also added that... good stuff.