Functional style with access to the timer  (#13)
* Functional style with access to the timer for s

* Review this PR
Removed some of the overloads, not needed any more if we can allow breaking changes (and we have)
This reduces the number of overloads quite a bit. We still need 4 for the cases with & without a return value, with and without async
Also, tests on changing the stat name

* Wording. Change "lambda" to "delegate" or "function" depending on if it has a return value
Since these are not necessarily lambdas (i.e. anon functions)

* I accidentally a letter