Tidy up IDisposables
Now that UdpTransport implements IDisposable, it needs to be tidied up so that it doesn't cause processes to crash when finalising instances when trying to tidy up the sockets.
This is particularly problematic on the finalizer thread, as ConcurrentBag<T> appears to use a ThreadLocal<T>, which can be disposed before the finalizer tries to tidy up the UdpTransport.
As a result, to stop IDisposable going viral, I've removed the constructor on StatsDPublisher that makes its own transport.
This means that this moves more and more into 4.0.0 territory, which is a shame.